moved the main subtopic component back into the subtopic file
